moving state up
